site/scripts/requirements.txt
cycler==0.10.0
functools32==3.2.3.post2
GDAL==2.1.0
matplotlib==2.0.0
numpy==1.11.3
olefile==0.44
Pillow==4.1.1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scipy==0.18.1
six==1.10.0
subprocess32==3.2.7
